# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Samir Portfolio API",
    "version": "1.0.0",
    "description": "API documentation for the Portfolio backend including admin, public pages, messages, translation, and activity audit trails."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Primary API Version 1"
    }
  ],
  "paths": {
    "/ping": {
      "get": {
        "summary": "Ping service health",
        "responses": {
          "200": {
            "description": "Ping OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin-login": {
      "post": {
        "summary": "Admin authentication login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["identifier", "password"],
                "properties": {
                  "identifier": { "type": "string", "description": "Username or Email address" },
                  "password": { "type": "string", "format": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "username": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/academics": {
      "get": {
        "summary": "Retrieve academic records list",
        "responses": {
          "200": {
            "description": "Academic records retrieved"
          }
        }
      },
      "post": {
        "summary": "Create academic record",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "institution": { "type": "string" },
                  "degree": { "type": "string" },
                  "start_year": { "type": "string" },
                  "end_year": { "type": "string" },
                  "logo_url": { "type": "string" },
                  "details_json": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Record created" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/translate": {
      "post": {
        "summary": "Batch translate texts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["texts", "language"],
                "properties": {
                  "texts": { "type": "array", "items": { "type": "string" } },
                  "language": { "type": "string", "example": "ko" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Translation complete" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}